#1047 P2 (step 3): extract session/wheel.rs (timer-wheel GC)
Third step of the session structural split. Lifts the bucketed
timer-wheel session GC subsystem (#965) out of session/mod.rs into
a new session/wheel.rs.

What moved into session/wheel.rs (80 LOC):
- pub(super) const WHEEL_BUCKETS, WHEEL_TICK_NS, FAR_FUTURE_OFFSET
- file-private const WHEEL_MASK + compile-time power-of-two assert
- pub(super) fn bucket_for_tick
- pub(super) fn target_tick_for
- pub(super) struct WheelEntry { pub(super) key, pub(super) scheduled_tick }
- pub(super) struct SessionWheel { pub(super) buckets, pub(super) cursor_tick, pub(super) initialized }
- impl SessionWheel { pub(super) fn new() -> Self }

Pure relocation, with visibility widened from file-private to
pub(super) for items that SessionTable in session/mod.rs accesses
directly: the struct fields (buckets, cursor_tick, initialized are
written/read from SessionTable methods like wheel_observe and
expire_stale_entries) and the helper fns/constants
(bucket_for_tick, target_tick_for, WHEEL_TICK_NS, FAR_FUTURE_OFFSET).

WHEEL_BUCKETS also widened to pub(super) so existing tests in
session/tests.rs that reference it directly still resolve. WHEEL_MASK
stays file-private since it's only used inside wheel.rs.

session/mod.rs adds:
    mod wheel;
    use wheel::{
        bucket_for_tick, target_tick_for, FAR_FUTURE_OFFSET, SessionWheel,
        WheelEntry, WHEEL_BUCKETS, WHEEL_TICK_NS,
    };

so SessionTable's existing call sites (e.g.
`self.wheel.buckets[bucket].push_back(WheelEntry { ... })`,
`now_ns / WHEEL_TICK_NS`, `bucket_for_tick(self.wheel.cursor_tick)`)
all continue to resolve unchanged.

LOC distribution after:
- session/mod.rs:   1,050 (was 1,113)
- session/wheel.rs:    80 (new)
- session/key.rs:     138 (unchanged)
- session/entry.rs:   132 (unchanged)
- session/tests.rs: 1,759 (unchanged)

cargo test --release: 865 passed, 0 failed.
